# Conflicts:
#	.idea/inspectionProfiles/Project_Default.xml

+++ app/src/main/java/com/uvg/mypokedex/data/model/Pokemon.kt
@@ -4,7 +4,7 @@ data class Pokemon(
     val id: Int,
     val type: String,
     val name: String,
-    val health: Double,
+    val health: Int,
 ) {
